requirements.txt
streamlit
streamlit-autorefresh
streamlit-authenticator
pandas
openpyxl
selenium
webdriver-manager
pyyaml
bcrypt
